Refactor frontend scripts into modules
