+++ index.html
@@ -159,7 +159,7 @@
     <script src="https://www.gstatic.com/firebasejs/9.22.2/firebase-app-compat.js"></script>
     <script src="https://www.gstatic.com/firebasejs/9.22.2/firebase-auth-compat.js"></script>
     <script src="https://www.gstatic.com/firebasejs/9.22.2/firebase-firestore-compat.js"></script>
-    <script type="module" src="script.js"></script>
+    <script type="module" src="assets/js/main.js"></script>
     <script>
       document.addEventListener("DOMContentLoaded", function() {
         const slides = [
